+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog for Pilau Base
 
 ## 2.2 (????-??-??)
-* 
+* `pilau_get_users_by_capability()`
 
 ## 2.1.3 (2015-08-19)
 * Added slug stopwords removal to compensate for Yoast SEO not working